[refactor] move kubernetes-related utility functions to kubernetes package
