+++ pkg/skaffold/kubernetes/util_test.go
@@ -1,0 +1,59 @@
+/*
+Copyright 2020 The Skaffold Authors
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package kubernetes
+
+import (
+	"testing"
+
+	"github.com/GoogleContainerTools/skaffold/testutil"
+)
+
+func TestSupportedKubernetesFormats(t *testing.T) {
+	tests := []struct {
+		description string
+		in          string
+		out         bool
+	}{
+		{
+			description: "yaml",
+			in:          "filename.yaml",
+			out:         true,
+		},
+		{
+			description: "yml",
+			in:          "filename.yml",
+			out:         true,
+		},
+		{
+			description: "json",
+			in:          "filename.json",
+			out:         true,
+		},
+		{
+			description: "txt",
+			in:          "filename.txt",
+			out:         false,
+		},
+	}
+	for _, test := range tests {
+		testutil.Run(t, test.description, func(t *testutil.T) {
+			actual := HasKubernetesFileExtension(test.in)
+
+			t.CheckDeepEqual(test.out, actual)
+		})
+	}
+}
