+++ pkg/skaffold/util/util.go
@@ -48,21 +48,6 @@ func RandomID() string {
 	return fmt.Sprintf("%x", b)
 }
 
-// These are the supported file formats for Kubernetes manifests
-var validSuffixes = []string{".yml", ".yaml", ".json"}
-
-// HasKubernetesFileExtension is for determining if a file under a glob pattern
-// is deployable file format. It makes no attempt to check whether or not the file
-// is actually deployable or has the correct contents.
-func HasKubernetesFileExtension(n string) bool {
-	for _, s := range validSuffixes {
-		if strings.HasSuffix(n, s) {
-			return true
-		}
-	}
-	return false
-}
-
 func StrSliceContains(sl []string, s string) bool {
 	return StrSliceIndex(sl, s) >= 0
 }
